Added ExcelFileManager, Update FitnessChecker
Moved read_file method to ExcelFileManager
